Extract class to reduce test code duplication
