+++ src/test/scala/com/asidatascience/configuration/DynamicConfigurationFromFileSpec.scala
@@ -2,46 +2,17 @@ package com.asidatascience.configuration
 
 import java.io.{File, PrintWriter}
 import java.nio.file.{Path, Paths}
-import java.util.concurrent.atomic.AtomicInteger
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.duration._
-import scala.util.{Success, Try}
+import scala.util.Try
 
-import akka.actor.ActorSystem
-import akka.testkit.TestKit
-import org.scalatest._
-import org.scalatest.concurrent.{Eventually, ScalaFutures}
+import org.scalatest.concurrent.Eventually
+import org.scalatest.Inside
 
 class DynamicConfigurationFromFileSpec
-extends TestKit(ActorSystem("dynamic-configuration-from-file-spec"))
-with FlatSpecLike
-with Matchers
-with BeforeAndAfterAll
-with ScalaFutures
+extends BaseSpec
 with Eventually
 with Inside {
-
-  override implicit val patienceConfig = PatienceConfig(
-    timeout = 5.seconds, interval = 50.millis)
-
-  override def afterAll(): Unit = {
-    TestKit.shutdownActorSystem(system)
-  }
-
-  case class Configuration(timestamp: Long)
-
-  private val dummyConfiguration = Configuration(1L)
-  private val dummyContents = "dummy-contents"
-
-  trait TestConfigurationParser {
-    val nHits = new AtomicInteger(0)
-
-    def parse(contents: String): Try[Configuration] = {
-      contents shouldEqual dummyContents
-      nHits.incrementAndGet()
-      Success(dummyConfiguration)
-    }
-  }
 
   private def withTemporaryFile(block: Path => Any): Any = {
     val file = File.createTempFile("dynamic-configuration", ".tmp")
@@ -60,13 +31,13 @@ with Inside {
 
   "DynamicConfigurationFromFile" should "return None initially" in
   withTemporaryFile { path =>
-    val parser = new TestConfigurationParser {}
+    val parser = new TestConfigurationParser(dummyContents)
     val configuration = newDynamicConfiguration(path, parser.parse)
     configuration.currentConfiguration shouldEqual None
   }
 
   it should "register an initial configuration" in withTemporaryFile { path =>
-    val parser = new TestConfigurationParser {}
+    val parser = new TestConfigurationParser(dummyContents)
     val configuration = newDynamicConfiguration(path, parser.parse)
 
     system.scheduler.scheduleOnce(1.second) {
